# example/openapi.json
{
  "components": {
    "schemas": {
      "CreateTodoReq": {
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "type": "object"
      },
      "DeleteTodoReq": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "DeleteTodoResp": {
        "properties": {},
        "type": "object"
      },
      "GetTodoReq": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "GetTodoResp": {
        "properties": {
          "todo": {
            "$ref": "#/components/schemas/Todo"
          }
        },
        "required": [
          "todo"
        ],
        "type": "object"
      },
      "ListTodosReq": {
        "properties": {},
        "type": "object"
      },
      "ListTodosResp": {
        "properties": {
          "todos": {
            "items": {
              "$ref": "#/components/schemas/Todo"
            },
            "type": "array"
          }
        },
        "required": [
          "todos"
        ],
        "type": "object"
      },
      "Todo": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "done",
          "created_at"
        ],
        "type": "object"
      },
      "UpdateTodoReq": {
        "properties": {
          "done": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "done"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/todos": {
      "get": {
        "description": "Returns an empty array if no todos exist.",
        "operationId": "getApiTodos",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListTodosResp"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Returns all todo items ordered by creation date."
      },
      "post": {
        "operationId": "postApiTodos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Creates a new todo item."
      }
    },
    "/api/todos/{id}": {
      "delete": {
        "operationId": "deleteApiTodosId",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Deletes a todo item by ID."
      },
      "get": {
        "description": "Returns 404 if the todo does not exist.",
        "operationId": "getApiTodosId",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTodoResp"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Retrieves a single todo item by ID."
      },
      "put": {
        "operationId": "putApiTodosId",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Updates the text and done status of a todo item."
      }
    }
  }
}